Define TypVar as a newtype-wrapped Any.
Previously, ANY and friends were defined as empty data types. This is
problematic because when doing x `seq` y, where x is a value of type
ANY, GHC notices that the seq must fail and doesn't generate any code
for y. When we then instantiate ANY with a non-empty type, things go
wrong.